delivery/requirements-dev.txt
black
flake8
flask-debugtoolbar
flask-shell-ipython
ipdb
ipython
isort
pytest
pytest-flask
pytest-cov
pylint-flask
flask-wtf
email_validator